Rename Tax Types Enum and add Barcode Types Enum

+++ Modules/Adjustment/database/migrations/2024_07_17_142919_CreateAdjustmentTable.php
@@ -14,8 +14,8 @@ return new class extends Migration
             $table->id();
             $table->date('date');
             $table->boolean('effected')->default(0);
-            $table->unsignedInteger('items');
-            $table->float('grand_total', 10, 4);
+            $table->unsignedTinyInteger('items');
+            $table->unsignedInteger('grand_total');
 
             $table->foreignId('user_id')->nullable()->constrained();
